starting to work on separated db drivers
Specific mongo DB code for the Email object extracted to a separate object which
is a static member of the Email object. Conversation, Userfilter, User, Domain and Config
will follow. This will allow me to implemente the SQLite and Postgres drivers with ease.

The DB driver will be selected with a version() at compile time so no unnecesary shit is
compiled.